requirements.txt
einops==0.2.0
imgaug==0.4.0
lmdb>=1.0.0
albumentations

gdown==4.4.0
scikit-image>=0.21.0
pillow==10.2.0
prefetch_generator==1.0.1
